only move unwrapExitCode
